refactors file system, and naming conventions. Deletes unused files
