+++ client/src/App.jsx
@@ -2,15 +2,14 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import PrivateRoute from './auth/PrivateRoute.jsx';
 
-import './App.css';
-
-import { Layout } from './components/Layout';
-import { Dashboard } from './components/dashboard';
-import { Onboard, SchoolDetailsForm } from './components/dashboard/onboarding';
+import Layout from './components/Layout';
+import Dashboard from './components/Dashboard';
+import Onboard from './components/Dashboard/Onboard';
+import SchoolDetailsForm from './components/Dashboard/Onboard/SchoolDetailsForm';
 
 function App() {
   return (
-    <div className="App">
+    <>
       <Layout>
         <Switch>
           <Route exact path="/" /> {/* Placeholder route - to be used later */}
@@ -22,7 +21,7 @@ function App() {
           <PrivateRoute path="/dashboard" component={Dashboard} />
         </Switch>
       </Layout>
-    </div>
+    </>
   );
 }
