+++ client/src/components/Dashboard/Onboard/SchoolDetailsForm.jsx
