--- client/src/components/Layout/Footer/index.js
@@ -1,1 +1,0 @@
-export { default as Footer } from './Footer';
