--- client/src/components/Layout/NavBar/index.js
@@ -1,1 +1,0 @@
-export { default as NavBar } from './NavBar';
