+++ client/src/components/Layout/index.jsx
@@ -1,14 +1,16 @@
 import React from 'react';
-import { NavBar } from './NavBar';
-import { Footer } from './Footer';
+import NavBar from './NavBar';
+import Footer from './Footer';
+
+// GlobalStyles will go here
 
 const Layout = ({ children }) => {
   return (
-    <div>
+    <>
       <NavBar />
       {children}
       <Footer />
-    </div>
+    </>
   );
 };
