--- client/src/components/dashboard/index.js
@@ -1,1 +1,0 @@
-export { default as Dashboard } from './Dashboard';
