--- client/src/components/dashboard/onboarding/index.js
@@ -1,2 +1,0 @@
-export { default as Onboard } from './Onboard.jsx';
-export { default as SchoolDetailsForm } from './SchoolDetailsForm.jsx';
